BACKUP_20250905_202309/requirements.txt
selenium>=4.27.0
requests>=2.32.5
beautifulsoup4>=4.12.3
google-generativeai>=0.8.5
python-dotenv>=1.0.1
pyyaml>=6.0.2
Pillow>=11.0.0
opencv-python>=4.10.0.84
numpy>=2.1.3
undetected-chromedriver>=3.5.5
schedule>=1.2.2
tqdm>=4.67.1
colorama>=0.4.6
